+++ paper/paper.md
@@ -1,0 +1,146 @@
+---
+title: "RydIQule Version 2: Enhancing graph-based modeling of Rydberg atoms"
+tags:
+  - Python
+  - Rydberg
+  - Atomic Physics
+  - Quantum Sensing
+  - Quantum
+  - Graph
+authors:
+  - name: Benjamin N. Miller
+    orcid: 0000-0003-0017-1355
+    equal-contrib: true
+    affiliation: 1 
+  - name: David H. Meyer
+    orcid: 0000-0003-2452-2017
+    equal-contrib: true
+    affiliation: 1
+  - name: Omar Nagib
+    orcid: 0000-0003-3345-4781
+    affiliation: 2
+  - name: Teemu Virtanen
+    orcid: 0000-0002-0653-2564
+    affiliation: 3
+  - name: Kevin C. Cox
+    orcid: 0000-0001-5049-3999
+    affiliation: 1
+affiliations:
+  - index: 1
+    name: DEVCOM Army Research Laboratory, 2800 Powder Mill Rd, Adelphi, MD, 20783, USA
+    ror: 011hc8f90
+  - index: 2
+    name: Department of Physics, University of Wisconsin-Madison, 1150 University Avenue, Madison, WI, 53706, USA
+    ror: 01y2jtd41
+  - index: 3
+    name: Naval Air Warfare Center, 1 Administration Circle, China Lake, CA, 93555, USA
+    ror: 01f0pxq13
+
+date: 13 March 2025
+bibliography: paper.bib
+aas-doi: 
+aas-journal:
+---
+# Summary
+<!--- This is an HTML comment in Markdown -->
+<!--- EACH SENTENCE SHOULD BE A NEWLINE --->
+Rydberg atomic radio-frequency (rf) sensors are an emerging technology platform that relies on vaporous atoms, interrogated with laser beams and nearly ionized, to receive rf signals.
+Rydberg rf sensors have a number of interesting fundamental distinctions from traditional receiver technologies,
+such as those based on metallic antennas,
+since they are governed by the quantum physics of atom-light interactions [@fancher_rydberg_2021].
+As Rydberg sensors quickly advance from laboratory experiments into fieldable devices,
+there is a need for a general software modeling tool that fully encompasses the internal physics of the sensor.
+The Rydberg Interactive Quantum Module (RydIQule) is a Python package designed to fill this need.
+
+RydIQule calculates the dynamics of atomic quantum states,
+with a particular emphasis on modeling thermal vapors of Rydberg atoms coupled to optical and radio-frequency electromagnetic fields.
+To accomplish this, a unique graph-based paradigm is used to represent the complex quantum system consisting of multiple energy levels and multiple electromagnetic fields,
+where each energy level is stored as a graph node and each electromagnetic coupling as a graph edge.
+RydIQule then generates a set of differential equations for the quantum state dynamics from the graph,
+using the Lindblad master equation formalism [@manzano_short_2020].
+Finally, RydIQule leverages linear equation solvers,
+such as those provided by NumPy [@harris_array_2020], SciPy [@virtanen_scipy_2020] or CyRK [@cyrk]
+to efficiently solve these systems and recover the quantum system response to arbitrary input fields.
+During the numerical solving, systems of equations are represented as tensor objects,
+allowing for efficient parameterization and computation of large sets of equations.
+All together, RydIQule provides a flexible platform for forward modeling Rydberg sensors while also providing a widely useful set of theoretical tools for fundamental exploration of atomic physics concepts.
+
+The initial public release of RydIQule in late 2023 built the core functionality described above [@miller_rydiqule_2024].
+Here we outline RydIQule's version 2 release which expands on its capabilities to more accurately model real-world atoms.
+
+# Statement of Need
+
[…]
