+++ .github/workflows/draft-joss.yml
@@ -16,7 +16,7 @@ jobs:
         uses: openjournals/openjournals-draft-action@master
         with:
           journal: joss
-          paper-path: paper/rq_v2_paper.md
+          paper-path: paper/paper.md
       - name: Upload Artifact
         uses: actions/upload-artifact@v4
         with:
