### 2025-04-02 #7988 @dihm
> please rename your paper to be `paper.md`

Should be fixed, I'll retrigger the pdf gen and see

> For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!

Fair enough. Thanks for getting us in the queue!
