### 2025-05-15 #7988 @phibeck
Hi @dihm could you elaborate on the developments of the package since the publication of the paper referenced in your repo, Computer Physics Communications, 294, 108952 (2024) https://doi.org/10.1016/j.cpc.2023.108952? Thanks
