### 2025-05-16 #7988 @dihm
@phibeck Certainly!

That publication corresponds to the v1.0.0 release of the package, which provided the core framework and tools that provide a flexible, fast, and computationally-efficient method for solving semi-classical master equations of atom-light quantum systems across large parameter spaces (with a particular focus on the needs of our original application in modeling thermal Rydberg sensors). Two levels of user interface classes are provided: `Sensor`, the physicist's low-level interface that provides full flexibility; and `Cell`, a higher level interface that enforces some basic assumptions and implements automations for modeling physical atoms.

While waiting for the review of that prior work to complete we began work on some of the outstanding deficiencies of v1. Put briefly, v2 expands the space of problems that can _easily and readily_ be modeled. The major efforts fall into three areas:

1. Initial uses for rydiqule were largely limited to modeling thermal Rydberg sensors, where a decent (and common) approximation is to completely ignore sublevel atomic structure. This greatly reduces the size of the atomic basis states that need to be considered. But this approximation breaks down fairly quickly even in thermal Rydberg sensors, much less other atomic systems. And while `Sensor` was fully flexible enough to handle manifolds of degenerate magnetic sublevels, in practice the manual book-keeping required by the user was arduous enough that it meant rydiqule was not useful for those types of models. To address this, we implemented structural labeling of graph nodes (ie tuple labels for the atomic states) so that groups of related atomic energy levels could be specified when creating the basis and when defining couplings between groups of levels. This necessitated a significant internal overhaul of `Sensor` to support the abstraction, particularly because we wanted to ensure backwards compatibility with v1 models.
2. Though `Cell` was intended to be a simplified and more automated interface of `Sensor` that required less detailed understanding of the underlying physics and calculations, it became clear quite quickly that implementation details and assumptions imposed by `Cell` in v1 made it difficult to use accurately. With v2, we performed a nearly complete re-write of `Cell` to improve its usability. In particular, we leveraged the improvements from point 1 to lower the number of required assumptions (ie proper atomic structure including magnetic sublevels are fully supported). We also significantly improved the interface to the underlying atomic properties calculation package (ARC), including upstreaming local extensions back to the ARC package.
3. A recent theoretical result (http://arxiv.org/abs/2501.06134) outlined a method that significantly improves the speed and accuracy of the Doppler averaged models. Given that the two above improvements will greatly increase system sizes and therefore extend solve times, we were eager to incorporate their result to support rydiqule's goal of fast, iterative modeling. Working with the lead author, we ported his Mathematica code to python and implemented them within rydiqule directly. Though presently an experimental feature in rydiqule since it only covers 1 dimension, we expect to extend to 3D in the next couple months and formalize the feature with the next minor release.

These efforts span from mid-2023 to Mar-2025, encompassing 953 commits over 98 pull requests from 5 contributors, with some being backported to v1 via the v1.1.0 and v1.2.x releases.

Ultimately, I realize that having a private development repo makes the review harder, but it is (unfortunately) a requirement of our employer. That said, I'm happy to provide access to whomever in the review process if they feel it would be helpful.
