### 2025-05-23 #7988 @kyleniemeyer
I agree that this is sufficient work towards a new major version that warrants JOSS review and publication.
