### 2025-06-09 481a37d
Update doppler_exact.py
Modified to implement changes in the Nagib/Walker paper
