### 2025-06-09 668a662
Update doppler_exact.py
fixed rho0 shape
