### 2025-06-12 160a79c
Update doppler_exact.py
made helper function _get_rho0 for finding the null vector. Changed averaging method to not require inverting matrix L.
