### 2025-06-12 ed5581e
Update doppler_exact.py
simplifying get_rho0 parameters and other small edits
