### 2025-06-13 2416df0
Update doppler_exact.py
fixed indexing error in _get_rho0
