### 2025-06-17 9f556b4
Update doppler_exact.py
changed variable names for clarity
