### 2025-06-18 265295e
Update doppler_exact.py
Changed to add hybrid doppler averaging that can handle any number of dimensions. Added necessary helper functions for doppler equations and slicing
