+++ src/rydiqule/doppler_exact.py
@@ -17,12 +17,16 @@ import warnings
 import numpy as np
 from importlib.metadata import version
 from copy import deepcopy
+import psutil
+from typing import Optional
 
 from scipy.special import erf
 
 from .sensor import Sensor
 from .sensor_utils import _hamiltonian_term, generate_eom, make_real, _squeeze_dims
 from .sensor_solution import Solution
+from .doppler_utils import *
+from .slicing.slicing import matrix_slice
 from .exceptions import RydiquleError, RydiquleWarning, PopulationNotConservedWarning
 
 
@@ -65,33 +69,247 @@ def _doppler_eigvec_array(lamdas: np.ndarray, rtol: float = 1e-5, atol: float = 
 
     return doppler_array
 
-
-def doppler_1d_exact(sensor: Sensor, rtol: float = 1e-5, atol: float = 1e-9) -> Solution:
+def _get_rho0(L0: np.ndarray) -> np.ndarray:
     """
-    Analytically solves a sensor in steady-state in the presence of 1 dimensional
-    Doppler broadening.
+    Helper function for computing rho0 (null vector) of a stack of equations of motion
+    via the inverse power method.
 
-    Uses the method outlined in Ref [1]_.
-    In particular, it uses Eq. 12 to analytically evaluate the Doppler average in 1D.
+    Paramaters
+    ----------
+    L0: np.ndarray
+        Stack of steady state Liouvillians
+    
+    Returns
+    -------
+    numpy.ndarray
+        Stack of steady state (vectorized) density matrices
+    """
 
-    This solver is considered more accurate than :func:`~.solve_steady_state`
+    stack_shape = L0.shape[0:-2]
+    n = int(np.sqrt(L0.shape[-1]))
+    rho0 = np.random.rand(*stack_shape, n**2)[..., np.newaxis]
+    rho0 /= np.linalg.norm(rho0, axis=-1, keepdims=True)
+
+    I = np.eye(n**2)
+    converged_flags = np.zeros(stack_shape, dtype=bool)
+        
+    # Compute rho0 by finding the null vector of L0 via the shifted inverse power method
+    for iteration in range(50):
+        if np.all(converged_flags):
+            break
+        
+        remaining_flags_index = np.where(~converged_flags)
+        
+        current_L0 = L0[remaining_flags_index]
+        current_shifted_L0 = current_L0 + 1e-14 * I
+
+        current_rho0 = rho0[remaining_flags_index]
+            
+        z = np.linalg.solve(current_shifted_L0, current_rho0) # compute (L0 + 1e-14)^-1 * rho0
+        rho0_new = z / (np.linalg.norm(z, axis=-2, keepdims=True) + np.finfo(float).eps)
+        
+        # Estimate magnitude of eigenvalues by Rayleigh quotient
+        L0rho0 = np.einsum('...ij,...j->...i', current_L0, rho0_new[..., 0])
+        numerator = np.sum(rho0_new[..., 0] * L0rho0, axis=-1)
+        denominator = np.sum(rho0_new[..., 0] * rho0_new[..., 0], axis=-1)
+        current_eigenvalues = np.abs(numerator / denominator)
+
+        rho_0_diff = np.linalg.norm(current_rho0 - rho0_new, axis=-1)
+
+        converged_flags[remaining_flags_index] = (rho_0_diff < 1e-15).any(axis=-1) | (current_eigenvalues < 1e-14)
+                
+        rho0[remaining_flags_index] = rho0_new
+
+    rho0 = rho0.squeeze(axis=-1)
+
+    rho0 *= np.sign(rho0[...,0])[...,None]  # remove arbitrary sign from null-vector so all pops are positive
[…]
