+++ src/rydiqule/doppler_exact.py
@@ -17,12 +17,16 @@ import warnings
 import numpy as np
 from importlib.metadata import version
 from copy import deepcopy
+import psutil
+from typing import Optional
 
 from scipy.special import erf
 
 from .sensor import Sensor
 from .sensor_utils import _hamiltonian_term, generate_eom, make_real, _squeeze_dims
 from .sensor_solution import Solution
+from .doppler_utils import *
+from .slicing.slicing import matrix_slice
 from .exceptions import RydiquleError, RydiquleWarning, PopulationNotConservedWarning
 
 
@@ -124,32 +128,188 @@ def _get_rho0(L0: np.ndarray) -> np.ndarray:
 
     return rho0
 
-def doppler_1d_exact(sensor: Sensor, rtol: float = 1e-5, atol: float = 1e-9) -> Solution:
+def get_doppler_equations_full_space(base_eoms: np.ndarray,
+                                     doppler_hamiltonians: np.ndarray,
+                                     velocities: np.ndarray) -> np.ndarray:
     """
-    Analytically solves a sensor in steady-state in the presence of 1 dimensional
-    Doppler broadening.
+    Constructs a stack of Doppler-shifted Liouvillians in the full n^2 space.
 
-    Uses the method outlined in Ref [1]_.
-    In particular, it uses Eq. 14 to analytically evaluate the Doppler average in 1D.
+    This is a custom version of the library's `get_doppler_equations` that
+    does NOT remove the ground state, making it compatible with the analytical solvers.
 
-    This solver is considered more accurate than :func:`~.solve_steady_state`
+    Parameters
+    ----------
+    base_eoms : numpy.ndarray
+        The base Liouvillian for a stationary atom (size n^2 x n^2).
+    doppler_hamiltonians : numpy.ndarray
+        Stack of Hamiltonians (size n x n) for the Doppler shifts.
+    velocities : numpy.ndarray
+        Mesh of velocity classes to sample.
+
+    Returns
+    -------
+    numpy.ndarray
+        A stack of Doppler-shifted Liouvillians (size n^2 x n^2).
+    """
+
+    n_squared = base_eoms.shape[-1]
+    obes_complex = _hamiltonian_term(doppler_hamiltonians)
+    dummy_const = np.zeros(n_squared)
+    base_doppler_shift_eoms = make_real(obes_complex, dummy_const, ground_removed=False)[0]
+    doppler_shift_eoms = np.tensordot(velocities, base_doppler_shift_eoms, ((0),(0)))
+
+    n_stacks = len(base_eoms.shape[:-2])
+    spatial_dim = base_doppler_shift_eoms.shape[0]
+    exp_dims = tuple(range(spatial_dim, spatial_dim+n_stacks))
+    doppler_eqns = np.expand_dims(base_eoms, 0) + np.expand_dims(doppler_shift_eoms, exp_dims)
+
+    return doppler_eqns
+
+def memory_size(shape: Tuple[int, ...], item_size: int) -> int:
+    """
+    Helper function that calculates the memory size of an array in bytes.
+    """
+    return np.prod(shape, dtype=np.ulonglong).item() * item_size
+
+def get_slice_num_hybrid(n: int, 
+                         param_stack_shape: Tuple[int, ...], 
+                         numeric_doppler_shape: Tuple[int, ...],
+                         n_slices: Optional[int] = None,
+                         debug: bool = False) -> Tuple[int, Tuple[int, ...]]:
+    """
+    Estimates memory and determines the number of slices for the HYBRID solver.
+
+    This version is tailored to the memory footprint of the hybrid algorithm,
+    which includes large intermediate arrays like the propagator and eigenvector stacks.
+    
+    Parameters
+    ----------
+    n : int
+        Size of the system basis.
[…]
