### 2025-06-23 7a68311
Update __init__.py
changed to doppler_hybrid which contains the 1d case
