### 2025-06-23 e92f06c
Update slicing.py
added get slice num function for the hybrid doppler method
