### 2025-06-23 4cda0c3
Moved utilities to appropriate files
