+++ src/rydiqule/doppler_exact.py
@@ -17,7 +17,6 @@ import warnings
 import numpy as np
 from importlib.metadata import version
 from copy import deepcopy
-import psutil
 from typing import Optional
 
 from scipy.special import erf
@@ -25,8 +24,8 @@ from scipy.special import erf
 from .sensor import Sensor
 from .sensor_utils import _hamiltonian_term, generate_eom, make_real, _squeeze_dims
 from .sensor_solution import Solution
-from .doppler_utils import *
-from .slicing.slicing import matrix_slice
+from .doppler_utils import doppler_classes, doppler_mesh, apply_doppler_weights
+from .slicing.slicing import matrix_slice, get_slice_num_hybrid
 from .exceptions import RydiquleError, RydiquleWarning, PopulationNotConservedWarning
 
 
@@ -165,104 +164,7 @@ def get_doppler_equations_full_space(base_eoms: np.ndarray,
 
     return doppler_eqns
 
-def memory_size(shape: Tuple[int, ...], item_size: int) -> int:
-    """
-    Helper function that calculates the memory size of an array in bytes.
-    """
-    return np.prod(shape, dtype=np.ulonglong).item() * item_size
-
-def get_slice_num_hybrid(n: int, 
-                         param_stack_shape: Tuple[int, ...], 
-                         numeric_doppler_shape: Tuple[int, ...],
-                         n_slices: Optional[int] = None,
-                         debug: bool = False) -> Tuple[int, Tuple[int, ...]]:
-    """
-    Estimates memory and determines the number of slices for the HYBRID solver.
-
-    This version is tailored to the memory footprint of the hybrid algorithm,
-    which includes large intermediate arrays like the propagator and eigenvector stacks.
-    
-    Parameters
-    ----------
-    n : int
-        Size of the system basis.
-    param_stack_shape : tuple of int
-        Tuple of sizes for the sensor's parameter axes (e.g., from L0.shape[:-2]).
-    numeric_doppler_shape : tuple of int
-        Tuple of sizes for the numeric doppler axes. Pass an empty tuple for the 1D case.
-    n_slices : int, optional
-        Manually override the minimum number of slices. If None, it's determined automatically.
-    debug : bool, optional
-        If True, prints detailed memory usage information.
-
-    Returns
-    -------
-    n_param_slices : int
-        Number of slices to use when iterating over the parameter stack.
-    out_sol_shape : tuple of int
-        Shape of the final, fully-solved solution array.
-    """
-    if n_slices is None:
-        n_slices = 1
-
-    # --- 1. Calculate Mandatory Memory (arrays that exist fully, regardless of slicing) ---
-    total_mem = psutil.virtual_memory().available
-    
-    # The full L0 matrix with all parameter stacks
-    l0_full_shape = (*param_stack_shape, n**2, n**2)
-    l0_full_mem = memory_size(l0_full_shape, 16) # complex128
-
-    # The final output solution array
-    out_sol_shape = (*param_stack_shape, n**2)
-    out_sol_mem = memory_size(out_sol_shape, 16) # Can be complex before final check
-    
-    mand_mem = l0_full_mem + out_sol_mem
-
-    # --- 2. Calculate Memory for a Single Slice ---
-    # This is the memory needed to process ONE parameter point through the hybrid algorithm.
-    
-    # Shape of the stacks over the numeric velocity grid
-    num_dop_stack_shape = (*numeric_doppler_shape, n**2, n**2)
-    num_dop_vec_shape = (*numeric_doppler_shape, n**2)
-
[…]
