### 2025-06-23 755e6c7
Update test_analytic_doppler.py
Edited to new hybrid function
