### 2025-06-24 4c8b419
Renamed solve_doppler_hybrid and moved helper function to doppler_utils
