### 2025-06-24 ebc2eb4
Update docs to include analytic method
