+++ docs/source/writeups/doppler.rst
@@ -1,7 +1,9 @@
 Doppler Averaging
 =================
 
-This document discusses the methods rydiqule uses to implement doppler-averaging of modelling results.
+This document discusses the methods rydiqule uses to implement doppler-averaging of modelling results. 
+As of version 2.1.0, rydiqule now provides two functions for implementing doppler-averaging: :func:`~.solve_steady_state` that averages numerically by sampling 
+and :func:`~.solve_doppler_hybrid` that averages one spatial dimension analytically and the remaining spatial dimensions numerically.
 The theoretical background is provided for completeness,
 but is fairly standard.
 Rydiqule's implementation of this is less obvious in order to optimize computational efficiency by fully leveraging numpy's vectorized operations.
@@ -19,11 +21,11 @@ Both assumptions greatly simplify doppler averaging.
 Specifically, these assumptions allow us to further assume that atoms in different velocity classes do not interact,
 meaning that doppler averaging entails a simple weighted average of the atomic response at different velocities by the velocity distribution.
 
-The basic process under these assumptions is as follows:
+.. The basic process under these assumptions is as follows:
 
-#. Calculate the density matrix solutions to the equations of motion for a wide range of velocity classes.
-#. Weight the density matrix solutions with the Maxwell-Distribution, assigning the relative atomic density per velocity class to each solution.
-#. Sum the weighted solutions.
+.. #. Calculate the density matrix solutions to the equations of motion for a wide range of velocity classes.
+.. #. Weight the density matrix solutions with the Maxwell-Distribution, assigning the relative atomic density per velocity class to each solution.
+.. #. Sum the weighted solutions.
 
 Note that rydiqule assumes a three dimensional distribution of velocities, as is the case for a vapor.
 
@@ -94,8 +96,8 @@ This allows us to readily produce the appropriate weighting distribution for 1, 
 without having to perform redundant calculations of distributions where the density matrices to be averaged do not depend on a particular :math:`v_i`.
 This function is implemented in :func:`gaussian3d`.
 
-Averaging Velocity Classes
---------------------------
+Numerically Averaging Velocity Classes
+--------------------------------------
 
 Given the above assumptions, the doppler average of the density matrix solutions is given by the integral
 
@@ -117,11 +119,11 @@ We again note that when all k-vectors along a particular axis are zero,
 :math:`\rho_{ij}(v_k, v_l, v_m)` is constant along that axis and that axis of the sum can be separated
 and assumed to sum to unity due to normalization of the weighting distribution along each dimension.
 
-Rydiqule's Implementation
--------------------------
+Rydiqule's Implementation (Numeric Method)
+++++++++++++++++++++++++++++++++++++++++++
 
 Rydiqule's implementation of Doppler averaging is optimized to minimize duplicate calculations and fully leverage numpy's vectorized and broadcasting operations.
-The general steps are are follows:
+The general steps of :func:`~.solve_steady_state` are as follows:
 
 #. Choose the doppler velocities to use for the mesh in the average.
[…]
