### 2025-06-26 675bb43
Add workflow job that ensures changelog is updated for each PR.
