### 2025-06-26 8dcc027
Fix changelog check to not run on PR merges
