+++ .github/workflows/checks.yml
@@ -69,7 +69,7 @@ jobs:
     # https://fredrikmeyer.net/2021/05/12/github-actions-changelog.html
     name: Changelog Updated
     runs-on: ubuntu-latest
-    if: github.repository_owner == 'QTC-UMD' || github.event_name == 'pull_request'
+    if: github.repository_owner == 'QTC-UMD' && github.event_name == 'pull_request'
     steps:
       - name: Checkout
         uses: actions/checkout@v4
