### 2025-06-26 d3b6050
changed erf to erfcx for numerical stability
