### 2025-06-26 7f06bee
Add 2D unit test to test_analytic_doppler.py
