### 2025-06-26 39c0e18
Update changelog.rst
