### 2025-06-27 #7988 @dihm
@phibeck Don't fret the delay. I understand that things happen. It actually works out well for us since I've had a summer student working on moving one of the new v2 features from experimental to fully supported (analytic doppler averaging in higher dimensions) and we expect to release v2.1 with the full support sometime next week. I am hoping it would be OK to include him as an author on the JOSS paper and update the paper to target version to v2.1 when that happens. Is that OK with you?

As for reviewers, I believe the following would be a good place to start: lamBOOO, PythonFZ, szhorvat, taladjidi, HugoStrand, a-eghrari
