### 2025-06-29 100934e
Tweak exception messages
