### 2025-06-29 0a3013f
Add unit tests that confirm input validation works for `solve_doppler_hybrid`
