### 2025-06-29 8955c71
Minor wording tweaks
