### 2025-06-30 2bfada7
Update changelog.rst
