### 2025-06-30 4c4997a
Rename `solve_doppler_hybrid` to `solve_doppler_analytic`
