### 2025-06-30 aebf14d
Swap a few instances of `hybrid` with `analytic` in the docstrings.
