### 2025-07-01 f77980b
Update sphinx version pins for RTD builds.
