+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=64", "wheel", "setuptools_scm>=8"]
+requires = ["setuptools>=64", "setuptools_scm>=8"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
@@ -54,9 +54,10 @@ backends = [
     "cyrk>=0.12.1",
 ]
 docs = [
-    "Sphinx==7.2.6",
+    "Sphinx==8.2.3",
     "furo==2024.8.6",
-    "myst-nb",
+    "myst-nb==1.2.0",
+    "myst-parser==4.0.1",
     "sphinx-copybutton",
     "sphinxext-opengraph",
     "sphinx-inline-tabs",
