### 2025-07-01 18c3d58
Update paper to target v2.1, which provides full support for analytic doppler solves.
Also adds a few newer citations for people using Rydiqule.
