+++ paper/paper.bib
@@ -132,6 +132,26 @@
 
 
 # science references
+@misc{cui_realizing_2025,
+      title={Realizing Quantum Wireless Sensing Without Extra Reference Sources: Architecture, Algorithm, and Sensitivity Maximization}, 
+      author={Mingyao Cui and Qunsong Zeng and Zhanwei Wang and Kaibin Huang},
+      year={2025},
+      eprint={2504.21234},
+      archivePrefix={arXiv},
+      primaryClass={cs.IT},
+      url={https://arxiv.org/abs/2504.21234}, 
+}
+
+@misc{glick_warm_2025,
+      title={Warm Vapor Rydberg EIT spectra in Doppler-free configurations}, 
+      author={Jeremy Glick and Brielle E. Anderson and T. Nathan Nunley and Josiah Bingaman and Jian Jun Liu and David H. Meyer and Paul D. Kunz},
+      year={2025},
+      eprint={2506.04504},
+      archivePrefix={arXiv},
+      primaryClass={physics.atom-ph},
+      url={https://arxiv.org/abs/2506.04504}, 
+}
+
 @misc{nagib_exact_2025,
   title = {Exact Steady State of Perturbed Open Quantum Systems},
   author = {Nagib, Omar and Walker, Thad G.},
