### 2025-07-01 #7988 @lamBOOO
> 👋 [@lamBOOO](https://github.com/lamBOOO), [@PythonFZ](https://github.com/PythonFZ) & [@taladjidi](https://github.com/taladjidi), would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @phibeck, I’m afraid I don’t have enough time right now, so I’ll have to decline.
