### 2025-07-01 6303b18
Fix block references
