### 2025-07-03 #7988 @PythonFZ
> 👋 [@lamBOOO](https://github.com/lamBOOO), [@PythonFZ](https://github.com/PythonFZ) & [@taladjidi](https://github.com/taladjidi), would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Unfortunately, I won't be able to make time for a review in the upcoming weeks.
