### 2025-07-03 #7988 @rkdan
> 👋 [@a-eghrari](https://github.com/a-eghrari), [@acrlakshman](https://github.com/acrlakshman) & [@rkdan](https://github.com/rkdan), would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Apologies, but I won't be able to review this work.
