### 2025-07-03 #7988 @a-eghrari
> :wave: @a-eghrari, @acrlakshman & @rkdan, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I am interested in reviewing this work.
