### 2025-07-04 #7988 @phibeck
Thank you very much @acrlakshman and @a-eghrari for agreeing to review! 🎉 Thanks also @PythonFZ and @rdkan for your quick responses - no problem.

I'll close the pre-review and start the review in another issue.
