### 2025-07-04 #8539 @editorialbot
[REVIEW]: RydIQule Version 2: Enhancing graph-based modeling of Rydberg atoms
**Submitting author:** @dihm (<a href="http://orcid.org/0000-0003-2452-2017">David H Meyer</a>)
**Repository:** https://github.com/QTC-UMD/rydiqule
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.1.2
