### 2025-07-04 f546c6e
Add missing DOIs and re-org bib file a small amount.
