+++ paper/paper.bib
@@ -80,7 +80,8 @@
   editor = {Varoquaux, Ga{\"e}l and Vaught, Travis and Millman, Jarrod},
   year = {2008},
   pages = {11--15},
-  address = {{Pasadena, CA USA}}
+  address = {{Pasadena, CA USA}},
+  doi = {10.25080/tcwv9851},
 }
 
 @article{virtanen_scipy_2020,
@@ -132,25 +133,6 @@
 
 
 # science references
-@misc{cui_realizing_2025,
-      title={Realizing Quantum Wireless Sensing Without Extra Reference Sources: Architecture, Algorithm, and Sensitivity Maximization}, 
-      author={Mingyao Cui and Qunsong Zeng and Zhanwei Wang and Kaibin Huang},
-      year={2025},
-      eprint={2504.21234},
-      archivePrefix={arXiv},
-      primaryClass={cs.IT},
-      url={https://arxiv.org/abs/2504.21234}, 
-}
-
-@misc{glick_warm_2025,
-      title={Warm Vapor Rydberg EIT spectra in Doppler-free configurations}, 
-      author={Jeremy Glick and Brielle E. Anderson and T. Nathan Nunley and Josiah Bingaman and Jian Jun Liu and David H. Meyer and Paul D. Kunz},
-      year={2025},
-      eprint={2506.04504},
-      archivePrefix={arXiv},
-      primaryClass={physics.atom-ph},
-      url={https://arxiv.org/abs/2506.04504}, 
-}
 
 @misc{nagib_exact_2025,
   title = {Exact Steady State of Perturbed Open Quantum Systems},
@@ -299,6 +281,28 @@
 
 
 # papers that have used rydiqule for modeling
+@misc{cui_realizing_2025,
+      title={Realizing Quantum Wireless Sensing Without Extra Reference Sources: Architecture, Algorithm, and Sensitivity Maximization}, 
+      author={Mingyao Cui and Qunsong Zeng and Zhanwei Wang and Kaibin Huang},
+      year={2025},
+      eprint={2504.21234},
+      archivePrefix={arXiv},
+      primaryClass={cs.IT},
+      url={https://arxiv.org/abs/2504.21234}, 
+      doi={10.48550/arXiv.2504.21234},
+}
+
+@misc{glick_warm_2025,
+      title={Warm Vapor Rydberg EIT spectra in Doppler-free configurations}, 
+      author={Jeremy Glick and Brielle E. Anderson and T. Nathan Nunley and Josiah Bingaman and Jian Jun Liu and David H. Meyer and Paul D. Kunz},
+      year={2025},
+      eprint={2506.04504},
+      archivePrefix={arXiv},
+      primaryClass={physics.atom-ph},
+      url={https://arxiv.org/abs/2506.04504}, 
+      doi={10.48550/arXiv.2506.04504},
+}
+
 @article{schlossberger_zeeman-resolved_2024,
   title = {Zeeman-Resolved {{Autler-Townes}} Splitting in {{Rydberg}} Atoms with Tunable Resonances and a Single Transition Dipole Moment},
   author = {Schlossberger, Noah and Rotunno, Andrew P. and {Artusio-Glimpse}, Alexandra B. and Prajapati, Nikunjkumar and Berweger, Samuel and Shylla, Dangka and Simons, Matthew T. and Holloway, Christopher L.},
