### 2025-07-04 e0c70f6
Add JOSS readme badge, update v1 Comp Phys Comms badge to make it distinct.
