### 2025-07-28 e378bf7
Decrease precision in ARC lifetime check comparison
