+++ src/rydiqule/cell.py
@@ -33,7 +33,7 @@ class Cell(Sensor):
     In addition to the core functionality of `~.Sensor`, this class requires labelling states
     with `namedtuple`s of quantum numbers, automatically calculating of state lifetimes and
     decoherences and tracking of of some physical laser parameters.
-    A key distictinction between a :class:`~.Cell` and a :class:`~.Sensor` is that
+    A key distinction between a :class:`~.Cell` and a :class:`~.Sensor` is that
     a cell supports (and requires) and absolute ordering of energy between states,
     which allows for implicit calculation of decay rates and transition frequencies.
 
@@ -55,7 +55,7 @@ class Cell(Sensor):
         atomic_states : list of A_QState
             List of :class:`~.atom_utils.A_QState` representing the states of the atom. 
             More details about the `A_QState` class can be found in its documentation, but it
-            includes the elemends (`n`, `l`, `j`, `m_j`, `f`, `m_f`). These represent the 
+            includes the elements (`n`, `l`, `j`, `m_j`, `f`, `m_f`). These represent the 
             usual Hydrogen-like atom quantum numbers with the usual restrictions:
                 
                 - `n` must be a positive integer.
@@ -75,7 +75,7 @@ class Cell(Sensor):
             they will be expanded into a list of states, with one corresponding to each value. If
             multiple quantum numbers are specified with lists, the resulting list of states will
             contain all combinations of values. Furthermore, the `j`, `m_j`, `f`, and `m_f` 
-            quantum numbers can each be specifed using `"all"`, which corresponds to a list of
+            quantum numbers can each be specified using `"all"`, which corresponds to a list of
             all physically allowed values for that quantum number. This convention allows quick
             specifications of entire manifolds of states to be added to the sensor. See the 
             `Examples` section to see how to use these powerful specifications.
@@ -88,22 +88,22 @@ class Cell(Sensor):
             on value of `beam_area`. See :meth:`~.Sensor.add_transit_broadening` for
             details on how transit broadening is treated. Default is None.
         gamma_mismatch : str or dict
-            How to resolve discrepancies between calculated eacb state lifetime and the sum of 
-            all transtion rates out of each state. In practice, these discrepancies are a result of
+            How to resolve discrepancies between calculated each state lifetime and the sum of 
+            all transition rates out of each state. In practice, these discrepancies are a result of
             transition pathways which exist between states when one is accounted for is `states`
             and one is not. For example, if there is a decoherent transition between states 3->1
             and states 3->2, but `states` only includes states 0, 1, and 3, the calculated 
             `gamma_lifetime` of state 3 will be greater than the sum of all computed
             `gamma_transition` values. In many cases, it is desirable to account for this
-            decoherence in other ways. The options for handling the dicrepancy are:
+            decoherence in other ways. The options for handling the discrepancy are:
 
                 - `"ground"` which adds a decoherent coupling coupling between a state `s` with
                   a discrepancy :math:`\\Delta \\gamma` and divides :math:`\\Delta \\gamma` among 
                   all the ground states (states matching the `n,l,j` values of the lowest energy
                   state). 
-                - `"all"` which divides :math:`\\Delta \\gamma` amongst all states in the 
-                  :class:`~.Cell` which already have a `"gamma_transition"` value. The fraction
-                  each transition gets is weighted by the fraction of the total thdat transitions
+                - `"all"` which divides :math:`\\Delta \\gamma` amongst all transitions in the 
+                  :class:`~.Cell` which already have a `"gamma_transition"` value from that state. The fraction
[…]

[…]
