### 2025-07-28 90bbd30
Fix precision in unit test
