+++ paper/paper.bib
@@ -50,19 +50,22 @@
   keywords = {Atom-light interaction,Electric field propagation,Faraday effect,Magneto-optics,Polarimetry,Spectroscopy,Stokes parameters,Voigt effect},
 }
 
-@misc{eckel_pylcp_2020,
-  title = {{{PyLCP}}: {{A}} Python Package for Computing Laser Cooling Physics},
+@article{eckel_pylcp_2022,
+  ids = {eckel_pylcp_2020},
+  title = {{{PyLCP}}: {{A Python}} Package for Computing Laser Cooling Physics},
   shorttitle = {{{PyLCP}}},
-  author = {Eckel, S. and Barker, D. S. and Norrgard, E. B. and Scherschligt, J.},
-  year = {2020},
-  month = nov,
-  number = {arXiv:2011.07979},
+  author = {Eckel, Stephen and Barker, Daniel S. and Norrgard, Eric B. and Scherschligt, Julia},
+  year = {2022},
+  month = jan,
+  journal = {Computer Physics Communications},
+  volume = {270},
   eprint = {2011.07979},
-  primaryclass = {physics, physics:quant-ph},
-  publisher = {{arXiv}},
-  doi = {10.48550/arXiv.2011.07979},
-  urldate = {2023-03-22},
-  keywords = {Physics - Atomic Physics,Physics - Computational Physics,Quantum Physics},
+  pages = {108166},
+  issn = {0010-4655},
+  doi = {10.1016/j.cpc.2021.108166},
+  urldate = {2025-07-29},
+  archiveprefix = {arXiv},
+  keywords = {Atomic physics,Laser cooling,Physics - Atomic Physics,Physics - Computational Physics,Python,Quantum Physics}
 }
 
 @misc{rochester_atomicdensitymatrix_2008,
@@ -134,17 +137,20 @@
 
 # science references
 
-@misc{nagib_exact_2025,
+@article{nagib_exact_2025,
   title = {Exact Steady State of Perturbed Open Quantum Systems},
-  author = {Nagib, Omar and Walker, Thad G.},
+  author = {Nagib, Omar and Walker, T. G.},
   year = {2025},
-  month = jan,
-  number = {arXiv:2501.06134},
+  month = jul,
+  journal = {Physical Review Research},
+  volume = {7},
+  number = {3},
   eprint = {2501.06134},
   primaryclass = {quant-ph},
-  publisher = {arXiv},
-  doi = {10.48550/arXiv.2501.06134},
-  urldate = {2025-01-28},
+  pages = {033076},
+  publisher = {American Physical Society},
+  doi = {10.1103/kgsg-3npp},
+  urldate = {2025-07-29},
   archiveprefix = {arXiv},
   keywords = {Quantum Physics}
 }
@@ -348,44 +354,49 @@
   keywords = {Physics - Atomic Physics,Quantum Physics}
 }
 
-@misc{su_two-photon_2024,
+@article{su_two-photon_2025,
   title = {Two-Photon {{Rydberg EIT}} Resonances in Non-Collinear Beam Configurations},
-  author = {Su, Kevin and Behary, Robert and Aubin, Seth and Mikhailov, Eugeniy and Novikova, Irina},
-  year = {2024},
-  month = dec,
-  number = {18035},
-  eprint = {18035},
-  publisher = {opticaopen},
-  doi = {10.1364/opticaopen.27941730.v1},
-  urldate = {2024-12-17},
-  archiveprefix = {opticaopen}
-}
-
-@misc{richardson_study_2023,
-  title = {Study of {{Angle}} of {{Arrival Estimation}} with {{Linear Arrays}} of {{Simulated Rydberg Atom Receivers}}},
[…]
