### 2025-07-29 271707e
Update `test_average_dipole_moment` to be compatible with ARC 3.9.0
