### 2025-07-31 e658011
Update changelog.rst
