### 2025-08-07 b940868
more doc updates
