+++ docs/source/changelog.rst
@@ -29,10 +29,10 @@ v2.1.0
 Improvements
 ++++++++++++
 
-- Fully support 1D/2D/3D doppler solves using the analytic-enabled solver, `solve_doppler_analytic`. The anlytic average
+- Fully support 1D/2D/3D doppler solves using the analytic-enabled solver, `solve_doppler_analytic`. The analytic average
   is applied to a single dimension, with other present dimensions treated numerically. This provides notable decreases in
   computation times and memory footprints and increases in solution accuracy in all spatial dimensions.
-- Added an example notebook, `Analyic Doppler Solver`, showcasing the new `solve_doppler_analytic`.
+- Added an example notebook, `Analytic Doppler Solver`, showcasing the new `solve_doppler_analytic`.
 - Extended documentation and unit unit-testing to include hybrid solver implementation.
 
 Deprecations
@@ -188,7 +188,7 @@ Improvements
 
 - Level diagrams now use `Sensor.get_rotating_frames` to provide better plotting of energy ordering of levels.
 - Level diagrams now allow for optional control of plotting parameters by manually specifying `ld_kw` options on nodes and edges.
-- Added the ability to specify energy level shifts (additional Hamiltonian digonal terms) not accounted for by the coupling infrastructure.
+- Added the ability to specify energy level shifts (additional Hamiltonian diagonal terms) not accounted for by the coupling infrastructure.
 
 
 Bug Fixes
@@ -210,8 +210,8 @@ v1.1.0
 Improvements
 ++++++++++++
 
-- Added the ability to specify hyperfine states in a `Cell`. They are distiguished by having 5 quantum numbers `[n, l, j, f, m_f]`.
-- `kappa` and `eta` are now proprties of `Cell` which are calculated on the fly.
+- Added the ability to specify hyperfine states in a `Cell`. They are distinguished by having 5 quantum numbers `[n, l, j, f, m_f]`.
+- `kappa` and `eta` are now properties of `Cell` which are calculated on the fly.
 - Separated rotating frame logic from hamiltonian diagonal generation into a new function `Sensor.get_rotating_frames()`.
   Allows for simple inspection of what rotating frame rydiqule is using in a solve.
 - Reworked the under-the-hood parameter zipping framework. This should have minimal impact on user-facing functionality.
@@ -264,7 +264,7 @@ Deprecations
   It is now `rabi_frequency` times the `time_dependence`.
 - Multiple sign errors have been corrected in `Sensor` and `Cell` with regards to detunings.
   Results that are asymmetric about zero detuning are likely to change.
-  Please ensure all couplings are following correct sign conventions for consisten results
+  Please ensure all couplings are following correct sign conventions for consistent results
   (ie second state of `states` tuple has higher energy).
 - most of the functions in experiments.py have been moved to become methods of `Solution` class.
 
@@ -327,7 +327,7 @@ Bug Fixes
 +++++++++
 
 - Fixed issue where solvers would save doppler axes labels and values even when they are summed over to the solution object
-- Fixed a bug where energy level diagrams broke when decochernce rates were scanned.
+- Fixed a bug where energy level diagrams broke when decoherence rates were scanned.
 - Fixed issue where compiled timesolvers could not solve doppler averaged problems.
 - Fixed issue where certain doppler solves could not be sliced correctly
 
@@ -386,16 +386,16 @@ Improvements
   
   - Gamma matrix is now calculated on the fly with the `decoherence__matrix()` method.
   - Decoherent transitions are now added with with the `add_decoherence()` function in `Sensor`.
-  - `Cell` now calculates tranistion frequencies and decay rates automatically and places them on the appropriate graph edges.
+  - `Cell` now calculates transition frequencies and decay rates automatically and places them on the appropriate graph edges.
 
[…]
