### 2025-08-08 76d8f2a
Lazy import leveldiagram only when `draw_diagram` is called.
Prevents one reason for matplotlib from being imported at rq import
