### 2025-08-08 d09dbad
First pass at lazy import of arc.
This, with other tweaks, removes arc and matplotlib from rq import
unless asked for by the user (ie `draw_diagram` or `Cell` used).

Works, but is a bit ugly. Reduces `rq` import from 2.8s to 1.5s.
