### 2025-08-11 #8539 @phibeck
Thanks @acrlakshman! Just to let you know, the JOSS review process tends to be more incremental than at other journals, so you’re welcome to submit sections of your review whenever it works best for you.
