### 2025-08-13 ae6c319
Rework arc atom lazy import to be simpler by using a constructor function instead of a dict of imported classes.
