+++ src/rydiqule/atom_utils.py
@@ -4,55 +4,44 @@ Utilities for interacting with atomic parameters and ARC.
 
 from scipy.constants import epsilon_0, hbar, c
 import numpy as np
-import re 
+import re
 from .sensor_utils import expand_statespec
 
-from typing import Union, Optional, Literal, Tuple, List, Dict, Callable, NamedTuple, Any
+from typing import Union, Optional, Literal, Tuple, List, Dict, Callable, NamedTuple, TYPE_CHECKING
 
-from .exceptions import RydiquleError
+from .exceptions import RydiquleError, AtomError
 
-__all__ = [
-    'QSpec',
-    'A_QState',
-    'QState',
-    'D1_excited',
-    'D1_states',
-    'D2_excited',
-    'D2_states',
-    'ground_state',
-    'expand_qnums',
-    'calc_eta',
-    'calc_kappa',
-    'validate_qnums',
-]
-__lazy_loads = [
-    'ATOMS'
-]
+if TYPE_CHECKING:
+    import arc.alkali_atom_data
 
-def __dir__() -> List[str]:
+ATOMS = {
+    'H': 'Hydrogen',
+    'Li6': 'Lithium6', 'Li7': 'Lithium7',
+    'Na': 'Sodium',
+    'K39': 'Potassium39', 'K40': 'Potassium40', 'K41': 'Potassium41',
+    'Rb85': 'Rubidium85', 'Rb87': 'Rubidium87',
+    'Cs': 'Caesium'
+}
+"""
+Alkali atoms defined by ARC that can be used with :class:`~.Cell`.
+"""
 
-    return sorted(__all__ + __lazy_loads)
+def _load_arc_atom(atom_flag: str) -> 'arc.alkali_atom_data.Alkali_Atom':
+    """
+    Function that lazy loads ARC atoms from :external+arc:module:`~arc.alkali_atom_data`
 
+    Returns
+    -------
+    arc.alkali_atom_data.Alkali_Atom
+        ARC alkali atom class associated with the provided atom_flag.
+    """
 
-def __getattr__(name: str) -> Any:
-
-    if name == 'ATOMS':
-        import arc.alkali_atom_data as arc_atoms
-
-        ATOMS = {
-            'H': arc_atoms.Hydrogen,
-            'Li6': arc_atoms.Lithium6, 'Li7': arc_atoms.Lithium7,
-            'Na': arc_atoms.Sodium,
-            'K39': arc_atoms.Potassium39, 'K40': arc_atoms.Potassium40, 'K41': arc_atoms.Potassium41,
-            'Rb85': arc_atoms.Rubidium85, 'Rb87': arc_atoms.Rubidium87,
-            'Cs': arc_atoms.Caesium
-        }
-        """
-        Alkali atoms defined by ARC that can be used with :class:`~.Cell`.
-        """
-        return ATOMS
-    else:
-        raise AttributeError(f'Module {__name__} does not have attribute {name}')
+    import arc.alkali_atom_data as arc_atoms
+    
+    if atom_flag not in ATOMS.keys():
+        raise AtomError(f"Atom flag must be one of {ATOMS.keys()}")
+    
+    return getattr(arc_atoms, ATOMS[atom_flag])() # instantiate the class here
 
 
 ground_n = {
