### 2025-08-15 526609f
Update changelog and fix lint.
