### 2025-08-18 0717745
Fix typo in type checking.
