### 2025-08-21 f1b5481
Fix Cell `test_decoherences` test so it actually checks if transit broadening is applied when provided by the constructor.
