### 2025-08-21 e63c6ad
Actually add transit broadening to `Cell`
