### 2025-08-21 505d41a
Drop use of numpy math functions when arrays are not present.
This improves typing a fair bit, and technically improves performance slightly.
