### 2025-08-21 c30935a
Fix `Cell.eta`, `Cell.kappa`, and `Cell.probe_freq` property caching to actually work.
Also adds some better type hinting.
