### 2025-08-21 0352bbd
Add unit tests for mismatch calculation with Cell NLJ and HFS states.
Note that both tests fail for mainline.
