### 2025-08-21 f5814a7
Fix gamma mismatch calculations to properly work for 'ground' with manifolds, 'all' with single states, and 'all' with manifolds.
