### 2025-08-22 d041fb4
Update doctest values to match updated Cell calculations.
