### 2025-08-29 c9699bc
Fix outdated mypy config reference
