### 2025-08-30 ccc4fab
Docs fixes for EOM page:
- Latex fix
- Add section describing how to undo computational basis transformation on solutions.
