### 2025-09-02 f23e5f5
Update links to exact steady state reference paper in docs.
