### 2025-09-04 9479a1a
Add `Solution.complex_rho` property to sensor solution for more ready conversion to the complex basis.
