### 2025-09-04 c6642f3
Fix error in `convert_complex_to_dm` where array would have complex dtype.
