### 2025-09-05 80f40d5
Bump default doppler width from 2\sigma to 2.2\sigma.
From comparisons with analytic solver, a slightly wider background doppler mesh
improves absolute normalization with sacrificing much is resolution.
