### 2025-09-05 d6556a8
Bump default doppler width to 2.5\sigma
