### 2025-09-05 7ac0b87
Expand Observables docs to describe version 2 semantics for observable calculation and how a user can calculate a custom observable.
