+++ docs/source/writeups/observables.rst
@@ -37,7 +37,7 @@ but written in terms of atomic constants :math:`\kappa` and :math:`\eta` (see [2
 We validate that rydiqule calculates these observable quantities in a manner consistent with
 a canonical reference, namely Steck's Quantum and Atom Optics notes [1].
 The unit tests in 
-`test_experiments.py <https://github.com/naqslab/Rydiqule/blob/master/tests/test_experiments.py>`_
+`test_experiments.py <https://github.com/qtc-umd/rydiqule/blob/main/tests/test_experiments.py>`_
 test that rydiqule and Steck align,
 but assume that the density matrix element :math:`\rho_{eg}` is correct.
 Validity of density matrix elements is checked in numerous other tests.
@@ -49,7 +49,7 @@ as is done in Steck's Quantum and Atom Optics notes, Eq. 5.273 [1].
 Another way to calculate the scattering rate is using the imaginary term of the susceptibility
 corresponding to the probing field coupling.
 This is the method Rydiqule uses, and can be found in Eq 6.73 of [1]. 
-This test is implemented in `test_experiments.py <https://github.com/naqslab/Rydiqule/blob/master/tests/test_experiments.py>`_
+This test is implemented in `test_experiments.py <https://github.com/qtc-umd/rydiqule/blob/main/tests/test_experiments.py>`_
 as the `test_OD_with_steck` unit test.
 
 Running just these observable and experiment unit tests can be done by installing the `pytest` dependencies (see :doc:`Unit Tests <../dev/tests>`),
@@ -57,7 +57,50 @@ and running the following command from the package parent directory.
 
 .. code-block:: shell
 
-    pytest .\tests -m experiments
+    pytest -m experiments
+
+User-Defined Observables
+------------------------
+
+With the release of version 2, rydiqule implements observable calculations in a more general way.
+This was done to properly support observable calculations for a coupling involving sublevels,
+where the calculated result relies on more than one matrix element.
+
+Any observable quantity :math:`\mathcal{O}` of a quantum state defined by a density matrix :math:`\rho`
+can be calculated using
+
+.. math::
+
+    \mathcal{O} = Tr(\rho A)
+
+where :math:`A` is the associated operator for the desired observable.
+This operation can be performed on any solution using the :meth:`.Solution.get_observable` method,
+where the user provides :math:`A` as a (generally complex) matrix the same shape as the system Hamiltonian.
+
+In rydiqule, we are typically concerned with the complex susceptibility of the medium :math:`chi`,
+whose associated operator is the dipole operator :math:`\hat{d}`.
+For a general system, this operator is broken into two parts:
+a scalar quantity :math:`d` (the dipole moment) that describes the overall strength of the coupling,
+and relative coupling strengths :math:`\hat{C}` that describe the relative strength of couplings between sublevels.
+These relative strengths are things like Clebsh-Gordon coefficients and are represented as a matrix.
+In rydiqule's implementation, these coefficients are the `coupling_coefficient` parameters.
+A more detailed description of how these are defined in `Cell`'s implementation is given in :doc:`sublevels`.
+This coefficient matrix can be obtained for any coupling using the :meth:`.Solution.coupling_coefficient_matrix` method.
+
+Finally, the :class:`.Solution` object has a method :meth:`~.Solution.coupling_coefficient_observable` which
+automatically obtains the coupling matrix and calculates the associated observable using the above functions.
+By default, it will use the probe coupling (i.e. first coupling added to a :class:`.Sensor`).
+This functionality is used under the hood to implement rydiqule's standard observable functions
+:meth:`~.Solution.get_susceptibility`, :meth:`~.Solution.get_OD`,
+:meth:`~.Solution.get_transmission_coef`, :meth:`~.Solution.get_phase_shift`.
+
[…]
