### 2025-09-13 e2dbd81
Update changelog for patch release
