+++ docs/source/changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Latest
+v2.1.1
 ------
 
 Improvements
@@ -31,10 +31,6 @@ Bug Fixes
   Also fix issue that prevented `gamma_mismatch='all'` from working if only 1 dephasing is present that needs modification.
 - Update RF heterodyne example notebook to use new `Cell` syntax.
 - Properly mark slow/high-memory analytic doppler test.
-
-Deprecations
-++++++++++++
-
 
 
 v2.1.0
