+++ docs/source/changelog.rst
@@ -1,6 +1,28 @@
 Changelog
 =========
 
+Latest
+------
+
+Improvements
+++++++++++++
+
+- Improved documentation based on JOSS reviewer @a-eghrari feedback.
+  Biggest additions include more detailed contribution guidelines and
+  a quantitative discussion of rydiqule performance (largely lifted from Comp Sci Comms article).
+- Added the `Solution.complex_rho` property which provides ready conversion of the solved
+  density matrices into the complex basis with ground state present.
+
+Bug Fixes
++++++++++
+
+
+
+Deprecations
+++++++++++++
+
+
+
 v2.1.1
 ------
 
