### 2025-09-13 c71bdc4
Fix error in `convert_complex_to_dm` where array would have complex dtype.
