### 2025-09-13 f190a45
Fix error in `convert_complex_to_dm` where array would have complex dtype.
