### 2025-09-23 05ddc4e
Add explicit contribution guide to docs, expanding on contribution guide in README.
