+++ docs/source/dev/contributing.rst
@@ -1,0 +1,133 @@
+Contributing
+============
+
+We actively encourage contributions and collaboration in the development of rydiqule.
+Unforunately, rydiqule's development is done privately, *for reasons*.
+If you would like to submit a PR for all but the most trivial of changes,
+please e-mail us directly (david.h.meyer3.civ@army.mil or kevin.c.cox29.civ@army.mil)
+so we can discuss the collaboration.
+
+Tips for a successful contribution
+++++++++++++++++++++++++++++++++++
+
+If you would like to submit a pull request that improves or fixes rydiqule,
+first off, thank you!
+A wider set of contributors to a project significantly improves the quality
+of the project as well as its pace of development.
+
+Secondly, because of limitations placed on rydiqule's primary developers,
+development deviates from a typical open-source project in that day-to-day work is done privately,
+with bulk public code releases.
+The most important difference for a potential contributor is that early communication
+with the core developers is even more important so we can advise how to move forward.
+
+Finally, writing code, especially a PR to someone else's project,
+is similar to writing a journal article for peer-review.
+The (code) reviewer needs to understand what you have done and why.
+Ultimately you need to *persuade* them to accept.
+Below is a general list of tips for producing a PR that can be merged quickly.
+
+Ask *before* doing a lot of work
+--------------------------------
+
+Unlike paper reviews, the code review process is not blind.
+As such, and as is common to any open-source project, it is good form to reach out to the developers
+*before* submitting a PR with significant changes, via an issue on github or a direct message to the devs.
+This can save you a lot of time as it allows the devs
+to provide high-level guidance on what approach is likely to work and/or be accepted.
+It avoids duplicated effort if we are already working on a solution.
+It avoids wasted effort on things we have already rejected as not viable.
+It also provides an opportunity for us to directly collaborate and help.
+Finally, it allows us to set expectations on scope,
+such as clarifying when a unit test isn't necessary.
+
+Make things digestible
+----------------------
+
+In much the same way a paper with multiple pages of non-stop equations can be hard to follow,
+submitting a PR with 1000s of changed lines of code, especially without having reached out first,
+makes the PR very difficult to digest.
+And we will not accept code we do not understand.
+
+The key to digestible PRs is breaking changes down into small, logical, digestible steps.
+
+- PRs are most like sections of a paper.
+  Ensure that each PR addresses a single task.
+  For example, don't implement unrelated features in a single PR,
+  or undertake large refactors alongside a new feature.
+  It is OK (even expected) to have a PR depend on another to be merged first.
+- Commits are most like paragraphs in a paper.
+  You should tend to use many small commits within a PR,
+  ensuring each commit has a single purpose and a sufficiently detailed commit message.
+  For example, moving a file and editing its contents should be separate commits.
+  Updating many inter-related type hints should be a single commit.
+- Careful application of git to edit commit history to organize edits after the fact
+  should be considered.
+  After all, it is rare the the order of discovery matches the best order for communicating the result.
+  But note that editing history *after* creating the PR leads to annoyance for
+  others who have already checked out your changes.
+- Finally, don't overthink it.
+  If the idea can be communicated in a concise single page,
+  it is generally best to keep it that way instead of fluffing out the manuscript.
+  The same goes for PRs: learn to recognize time-wasting bloat and low-value perfection chasing.
+  If the change is straight-forward, don't waste time checking every box or revising commit messages.
+
+Follow style guides
+-------------------
+
+Submitting a PR that does not follow standard conventions
+makes it much harder for the reviewer to follow.
+Even if the underlying work is fundamentally correct,
+the differences in (ultimately arbitrary) styles consumes the bulk of the review effort.
+
[…]
