### 2025-10-31 e7388e0
Move dev and docs optional dependencies to dependency-groups.
