### 2025-10-31 884ce88
Update RTD config to install from dependency groups
