+++ readthedocs.yaml
@@ -9,6 +9,11 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.11"
+  jobs:
+    install:
+      # since install step is overridden, pip not updated automatically
+      - pip install --upgrade pip
+      - pip install --group 'docs'
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
@@ -20,11 +25,3 @@ sphinx:
 formats:
   - pdf
   - epub
-
-# Optionally set the version of Python and requirements required to build your docs
-python:
-   install:
-      - method: pip
-        path: .
-        extra_requirements:
-            - docs
