### 2025-10-31 465e58f
Make project compatible with uv.
Ensures docs dependency group uses python>=3.11
and that the uv.lock file is ignored.
