### 2025-10-31 31c78ea
Ensure package is actually installed.
