+++ readthedocs.yaml
@@ -13,6 +13,7 @@ build:
     install:
       # since install step is overridden, pip not updated automatically
       - pip install --upgrade pip
+      - pip install .
       - pip install --group 'docs'
 
 # Build documentation in the docs/ directory with Sphinx
