### 2025-11-19 e017fa6
Fix bug where Cell.kappa/eta errors if the first edge in the manifold is not dipole allowed.
