### 2025-11-20 #8539 @acrlakshman
> Hi [@phibeck](https://github.com/phibeck) is there a path forward here? We are now past four months on the review and it would be nice wrap this up for the sake of the student coauthors' fellowship applications going in now.

Looks like I am one of the bottlenecks here. Let me finish my review this weekend and update.

@phibeck
