### 2025-11-20 8243d96
Also prevent changelog checks if the PR title contains 'WIP' or is a draft.
