### 2025-11-20 4da25b4
Add source link for label-based change.
Mostly to test if check fails correctly.
