### 2025-11-20 e8cccca
Ensure numpy v1 unit test actually runs against numpy v1
