### 2025-12-23 c6cd9de
Add `ipykernel` and `tests` to the dev dependency.
This ensures uv auto installs them when working on rydiqule directly.
