+++ pyproject.toml
@@ -69,6 +69,9 @@ tests = [
     "coverage",
     "pytest-cov",
 ]
+profiling = [
+    "line_profiler==3.5.1", # pin to version that supports cython line profiling
+]
 mypy = [
     "mypy",
     "numpy>=1.21",
@@ -77,10 +80,17 @@ mypy = [
     "types-psutil",
 ]
 linter = ["ruff"]
-dev = ["setuptools-scm"]
+dev = [
+    "setuptools-scm",
+    "ipykernel",
+    {include-group = "tests"},
+]
 
 [tool.uv.dependency-groups.docs]
 requires-python = ">=3.11"
+
+[tool.uv.dependency-groups.profiling]
+requires-python = "<3.11"
 
 [tool.ruff]
 line-length = 100
