### 2025-12-23 f453b1f
Update installation docs to cover using `uv`
