### 2026-01-14 #8539 @phibeck
Hi @dihm do you have suggestions for another reviewer? It could be someone in the field or even an active user as long as there is no conflict of interest:

> As a reviewer (or editor), COIs are your present or previous association with any authors of a submission: recent (past four years) collaborators in funded research or work that is published; and lifetime for the family members, business partners, and thesis student/advisor or mentor. In addition, your recent (past year) association with the same organization of a submitter is a COI, for example, being employed at the same institution.
