### 2026-01-17 #8539 @taladjidi
@phibeck  seems I missed the shot on this one sorry ! But keep me posted, I currently have some availability for reviewing for JOSS 😄  cheers !
