### 2026-01-17 #8539 @arfon
@phibeck @nikolasibalic – apologies, Editorialbot gave you the wrong checklist, I've updated it to match the pre-2026 submission criteria for JOSS.
