### 2026-01-18 #8539 @phibeck
> [@phibeck](https://github.com/phibeck) seems I missed the shot on this one sorry ! But keep me posted, I currently have some availability for reviewing for JOSS 😄 cheers !

Hi @taladjidi no worries, and thanks for letting me know. I may take you up on this offer. :)
