### 2026-01-18 #8539 @nikolasibalic
# Review

In the software library RydIQUle, presented by Benjamin N. Milelr et al., a suite of tools is provided for calculating the response of Rydberg atomic sensors to RF/THz electromagnetic radiation. The updated version 2 improves both the user interface for handling complex atom structures and automates calculations for complex real-world examples. Additionally, it improves Doppler averaging calculations.

The manuscript is clearly written. The documentation is exhaustive, covering in clear language new user onboarding and installation, physics details, API details, and a number of end-to-end use case Jupyter notebooks that reproduce results from the literature. Reasonable amount of automated tests is also included. Helpful instructions for contributors and detailed release notes for each release are also included. It follows open-source software (OSS) practices.

The software has already been used in several research papers cited in the paper. It is of significance to the wide community of physicists in ongoing research in new Rydberg-based atomic sensors. I recommend it for publication in JOSS once the authors have a chance to address two minor issues highlighted below.

# Minor issues

- [x] **Analytics Doppler Averaging**

I found the claim about memory footprint, which grows when doing Doppler averaging, a bit confusing. While it does take time to sample and solve for each combination of atom velocities, there is no algorithmic need to fit all calculations in memory at the same time since each combination of sampled velocities can be solved independently. Could the authors please clarify this statement? If I am right about the essential decoupling of calculations, I believe the memory improvement focus should be left out of the discussion, as it's more reflective of the difference between two implementations, arguably not changing the state-of-the-art for memory usage, where sequential solving of decoupled equations wins. Discussion of analytic speedup, on the other hand, is valid.

- [x] **Example notebooks**
There were a few issues running the example notebooks, which I separately reported as [Issue](https://github.com/QTC-UMD/rydiqule/issues/58).

@phibeck @dihm
