### 2026-01-21 #8539 @dihm
@nikolasibalic I appreciate your thorough and prompt review and I'm glad the project has made a favorable impression. I've addressed the two minor issues via qtc-umd/rydiqule#59 and qtc-umd/rydiqule#60, and provide extra background for the changes below.

> I found the claim about memory footprint, which grows when doing Doppler averaging, a bit confusing. While it does take time to sample and solve for each combination of atom velocities, there is no algorithmic need to fit all calculations in memory at the same time since each combination of sampled velocities can be solved independently. Could the authors please clarify this statement? If I am right about the essential decoupling of calculations, I believe the memory improvement focus should be left out of the discussion, as it's more reflective of the difference between two implementations, arguably not changing the state-of-the-art for memory usage, where sequential solving of decoupled equations wins. Discussion of analytic speedup, on the other hand, is valid.

This is a subtle point, and you are of course correct. I do have one small caveat to add though. When accounting for the limitations of the Python interpreter, fitting all the calculations into memory becomes an pseudo-algorithmic requirement for us. Since each parameter set is independent, they _are_ solved independently (at least for steady-state). But if we implement that loop in python, the interpreter overhead completely dwarfs the computational time by orders of magnitude. To achieve the performance we have, we needed to move that loop down into C, which generally requires allocating as many sets of equation as our memory allows, and passing the sets to `numpy.linalg.solve` so it can handle the C-looping and array broadcasting efficiently. Again, you are right that algorithmically this is not a requirement, but implementing that space-efficient version of the algorithm in python with the tools available to us would lead to significant performance hits thanks to the realities of the python interpreter. This non-fundamental but eminently practical consideration is covered in the Comp.Phys.Comms paper, though that discussion did not make it into the docs themselves. To help address this concern, we've included it in the Intro notebook under the memory limitations section.

There is a purely implementation detail aspect that directly plays into your point, however. We have machinery in place that allows for python-looping to relax the memory constraints of our implementation. But we also special case the handling of Doppler averaging (relative to other experimental parameters), making it so Doppler averaging cannot leverage that framework. The result is that our implementation requires the entire Doppler-average to fit in memory to solve. This aspect is entirely due to a design choice on our part, which we hope to one day correct. If we did support this, then the required minimum memory footprints of solves between the analytic-hybrid and pure numeric approaches would be basically identical (though, again, using that minimum memory amount would significantly slow down the pure numeric part of the solve).

To fully answer this concern, we have made a minor change to the JOSS paper (qtc-umd/rydiqule@22d4bca6d164152e9f696edf8a17cb260f1db7ed) with the sentence mentioning memory now reading "This reduction in dimensionality results in over an order of magnitude reduction in calculation time while returning significantly higher accuracy solutions, with the added benefit of lower memory footprint over RydIQule's current implementation." We have also updated the Analytic Doppler Demo notebook memory comparisons to use the minimum required memory footprint instead of the full (which ends up being the same fraction in 2D/3D due to the implementation detail described above). I'll note that in that notebook, we have been careful to state that comparisons are only between our two implementations since the audience/purpose for that notebook is more existing users of rydiqule (to encourage them to use the new method) rather than justifying any state-of-the-art claims, which we leave to the more science-y references (Nagib et al in this case). If there are other places you noticed in the docs make the erroneous implied claim, please point them out so I can correct them as well.

> There were a few issues running the example notebooks...

This is a perennial problem with this project. Since we don't have automated testing of the notebooks themselves, they tend to slip relative to the implementation and our manual attempts to catch those things seems to always be lacking. Thank you for identifying more of these issues so we could stamp them out.
