### 2026-01-22 #8539 @dihm
@phibeck The new release target is now 2.1.2. The Zenodo archive DOI is [10.5281/zenodo.18341352](https://doi.org/10.5281/zenodo.18341352)

I have confirmed accuracy of titles, authors, and ORCiDs in the paper.
