### 2026-01-27 #8539 @phibeck
@dihm thanks for your patience. Here are my comments for the manuscript. Please have a look when you have a moment.

- lines 5-9: in your affiliations, please spell out USA
- line 11: Is it common to write the "rf" acronym in lowercase letters? It seems unusual, but I can see the literature is mixed
- line 44: I'd recommend "treated on average" or "averaged over"
- line 60: "codebase"
- line 68: "Its internals"
- line 78: "labeling"
- line 89: remove [] around the citation for author-in-text citation
- line 97: I realize this latest statement is a bit unclear to me. Could you please rephrase the statement without using "current" since it will become outdated soon?

References:
- line 144: please capitalize vapor, Rydberg, Doppler
