### 2026-01-27 #8539 @dihm
@phibeck I've made the suggested changes. Three clarifying comments below:

- line 11: using "rf" as the acronym comes from the style guide for Physical Review journals (unfortunately) so we tend to lean that way.
- line 97: This sentence was modified for reviewer 2 and I clearly tried doing too much. I've expanded the context and broken it into two sentences to hopefully make it clearer. I also changed "current" to "original".
- line 144: Turns out this reference has since published, so I updated to the current title, which is slightly different and has correct capitalization.
