### 2026-01-27 #8539 @phibeck
Thanks @dihm for taking care of this! Would you mind also making these last two suggested changes, then we're ready to send this off.

- line 78: "labeling"
- line 89: remove [] around the citation for author-in-text citation
