### 2026-01-27 #8539 @phibeck
Hi @dihm could you remove one of the occurrences of `harris_array_2020` in the `paper.bib` file? Thanks.
