### 2026-01-27 #8539 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
